Fix sleep-control bug (extra indent) & Remove extra events (stop & resume) & Fix float error in pass_encoder: secrets.token_hex(int)

+++ pass_encoder/encoder.py
@@ -19,7 +19,7 @@ menu = """
 # functions
 
 def generate_pass():
-    return secrets.token_hex(PASS_LENGTH / 2);
+    return secrets.token_hex(int(PASS_LENGTH / 2));
 
 def change_pass(user_name, password):
     args = [
